coainventory/requirements.txt
django-simple-history
django-crispy-forms